Hotfix/bug fixes

+++ fair/register.py
@@ -88,7 +88,7 @@ def fetch_registrations(
     typing.List[str]
         list of registered object URLs
     """
-    _expected_keys = [
+    _expected_keys_external_object = [
         "root",
         "path",
         "file_type",
@@ -97,9 +97,24 @@ def fetch_registrations(
         "public",
     ]
 
+    _expected_keys_data_product = [
+        "root",
+        "path",
+        "file_type",
+        "version",
+        "public"
+    ]
+
     _stored_objects: typing.List[str] = []
 
     for entry in user_config_register:
+        logger.debug(f"entry {entry}")
+
+        if "external_object" in entry:
+            _expected_keys = _expected_keys_external_object
+        else:
+            _expected_keys = _expected_keys_data_product
+
         for key in _expected_keys:
             if key not in entry and key not in entry["use"]:
                 raise fdp_exc.UserConfigError(
@@ -166,18 +181,22 @@ def fetch_registrations(
         _search_data["version"] = entry["use"]["version"]
         _namespace = entry["use"]["namespace"]
 
-        if not _identifier and not _unique_name:
-            raise fdp_exc.UserConfigError(
-                "Expected either 'unique_name' or 'identifier' in 'register' item"
-            )
+        if _external_object:
+            if not _identifier and not _unique_name:
+                raise fdp_exc.UserConfigError(
+                    "Expected either 'unique_name' or 'identifier' in 'register' item"
+                )
 
-        elif _identifier and _unique_name:
-            raise fdp_exc.UserConfigError(
-                "Only one unique identifier may be provided (doi/unique_name)"
-            )
-
+            elif _identifier and _unique_name:
+                raise fdp_exc.UserConfigError(
+                    "Only one unique identifier may be provided (doi/unique_name)"
+                )
+        # Set Remove to True by default so the tempory file gets deleted
+        _remove = True
         if "cache" in entry:
             _temp_data_file = entry["cache"]
+            # Don't delete the tempory file if it's from a cache
+            _remove = False
         else:
             _local_parsed = urllib.parse.urlparse(local_uri)
             _local_url = f"{_local_parsed.scheme}://{_local_parsed.netloc}"
@@ -207,7 +226,8 @@ def fetch_registrations(
                 " present with this name, deleting temporary data file",
                 _name,
             )
-            os.remove(_temp_data_file)
+            if _remove:
+                os.remove(_temp_data_file)
             continue
 
         # Item found but not hash matched retrieve a version number
@@ -239,7 +259,8 @@ def fetch_registrations(
         logger.debug("Saving data file to '%s'", _local_file)
         shutil.copy(_temp_data_file, _local_file)
 
-        os.remove(_temp_data_file)
+        if _remove:
+            os.remove(_temp_data_file)
 
         if "public" in entry:
             _public = entry["public"]
